restructure findr adapter repo
