+++ src/adapters/http.ts
@@ -2,10 +2,10 @@
 
 import express, { Request, Response, Express } from 'express';
 import bodyParser from 'body-parser';
-import { LocalMessageSaver } from './message-save-local.js';
-import { extractYearMonthDate } from './extract-year-month-date.js';
-import { FileNameUtility } from './filename-utility.js';
-import { DataTransferHandler } from './storage-index.js';
+import { LocalMessageSaver } from '../utilities/message-save-local.js';
+import { extractYearMonthDate } from '../utilities/extract-year-month-date.js';
+import { FileNameUtility } from '../utilities/filename-utility.js';
+import { DataTransferHandler } from '../storage-index.js';
 
 interface HttpAdapterConfig {
     source: string;
